requirements.txt
### 5. requirements.txt

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# ML/AI
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=1.7.0
opencv-python>=4.8.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Maps (Phase 3)
pydeck>=0.8.0
folium>=0.14.0
streamlit-folium>=0.13.0
geopy>=2.4.0

# Authentication (Phase 2)
bcrypt>=4.0.0
python-jose>=3.3.0
passlib>=1.7.4

# QR Code
qrcode>=7.4.0
python-barcode>=0.15.0

# SMS/Notifications (Optional)
twilio>=8.9.0

# Environment
python-dotenv>=1.0.0

# Utils
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0